Добавлена проверка прав.

+++ backend/controllers/AreaNodeController.php
@@ -2,6 +2,7 @@
 
 namespace backend\controllers;
 
+use common\models\User;
 use Throwable;
 use Yii;
 use common\models\AreaNode;
@@ -84,6 +85,10 @@ class AreaNodeController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can(User::PERMISSION_ADMIN)) {
+            return $this->redirect('/site/index');
+        }
+
         $model = new AreaNode();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -104,6 +109,10 @@ class AreaNodeController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can(User::PERMISSION_ADMIN)) {
+            return $this->redirect('/site/index');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -126,6 +135,10 @@ class AreaNodeController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can(User::PERMISSION_ADMIN)) {
+            return $this->redirect('/site/index');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
